src/multimodal_modal/backend/requirements.txt
# requirements.txt

fastapi
uvicorn
google-cloud-texttospeech
openai
requests
webrtcvad
gradio
f5-tts
python-dotenv